requirements.txt
langchain>=0.1.0
langchain_community>=0.0.30
python-dotenv>=1.0.0
requests>=2.31.0
